requirements.txt
python-telegram-bot
python-dotenv
requests
HTTPServer
fastapi
flask[async]